applications/backend/pyproject.toml
[tool.poetry]
name = "moodit"
version = "0.1.0"
description = ""
authors = ["sethantanah <sethsyd32@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
pydantic = "^2.7.4"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
python-jose = "^3.3.0"
pydantic-settings = "^2.3.3"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
